+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Mon Feb 11 17:59:20 UTC 2019
+
+* Add deviceFreeSpace to video operative events [#8179](https://github.com/Applifier/unity-ads-webview/pull/8179)
+* Change text of the Delete your data button [#8180](https://github.com/Applifier/unity-ads-webview/pull/8180)
+* Add list of 3rd parties to the new privacy and consent views [#8177](https://github.com/Applifier/unity-ads-webview/pull/8177)
+* Renamed NewVideoOverlay -> VideoOverlay, new-video-overlay -> video-overlay [#8152](https://github.com/Applifier/unity-ads-webview/pull/8152)
+* Add close button to the new Privacy settigns view [#8106](https://github.com/Applifier/unity-ads-webview/pull/8106)
+
 # Fri Feb  8 20:37:59 UTC 2019
 
 * Install button experiment for rewarded ad videos [#8055](https://github.com/Applifier/unity-ads-webview/pull/8055)